rename codec package to scale
