+++ dot/network/message.go
@@ -23,9 +23,9 @@ import (
 	"math/big"
 
 	"github.com/ChainSafe/gossamer/dot/core/types"
-	scale "github.com/ChainSafe/gossamer/lib/codec"
 	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/common/optional"
+	"github.com/ChainSafe/gossamer/lib/scale"
 )
 
 //nolint
