+++ lib/babe/runtime.go
@@ -18,8 +18,8 @@ package babe
 
 import (
 	"github.com/ChainSafe/gossamer/dot/core/types"
-	scale "github.com/ChainSafe/gossamer/lib/codec"
 	"github.com/ChainSafe/gossamer/lib/runtime"
+	"github.com/ChainSafe/gossamer/lib/scale"
 
 	log "github.com/ChainSafe/log15"
 )
