+++ lib/scale/helpers_test.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
 
-package codec
+package scale
 
 // make byte array with length specified; used to test byte array encoding
 func byteArray(length int) []byte {
